:recycle: Refactor the code to remove the cyclic dependency in InternedTerm
