+++ grisette.cabal
@@ -99,19 +99,17 @@ library
       Grisette.Internal.IR.SymPrim
       Grisette.IR.SymPrim
       Grisette.IR.SymPrim.Data.IntBitwidth
+      Grisette.IR.SymPrim.Data.Prim.GeneralFun
       Grisette.IR.SymPrim.Data.Prim.Helpers
       Grisette.IR.SymPrim.Data.Prim.InternedTerm.Caches
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.InternedCtors
       Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm
       Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-      Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermSubstitution
       Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermUtils
       Grisette.IR.SymPrim.Data.Prim.Model
       Grisette.IR.SymPrim.Data.Prim.ModelValue
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Bits
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
       Grisette.IR.SymPrim.Data.Prim.PartialEval.BV
-      Grisette.IR.SymPrim.Data.Prim.PartialEval.GeneralFun
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Integral
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Num
       Grisette.IR.SymPrim.Data.Prim.PartialEval.PartialEval
