+++ src/Grisette/Core/Data/Class/SubstituteSym.hs
@@ -51,12 +51,12 @@ import Generics.Deriving
   )
 import Generics.Deriving.Instances ()
 import Grisette.Core.Data.BV (IntN, WordN)
+import Grisette.IR.SymPrim.Data.Prim.GeneralFun (substTerm)
 import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
   ( LinkedRep (underlyingTerm),
     SupportedPrim,
     TypedSymbol,
   )
-import Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermSubstitution (substTerm)
 import Grisette.IR.SymPrim.Data.SymPrim
   ( SymBool (SymBool),
     SymIntN (SymIntN),
