+++ src/Grisette/Core/Data/Class/ToSym.hs
@@ -59,10 +59,10 @@ import Grisette.Core.Data.BV
   )
 import Grisette.Core.Data.Class.Solvable (Solvable (con))
 import Grisette.IR.SymPrim.Data.IntBitwidth (intBitwidthQ)
+import Grisette.IR.SymPrim.Data.Prim.GeneralFun (type (-->))
 import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
   ( LinkedRep,
     SupportedPrim,
-    type (-->),
   )
 import Grisette.IR.SymPrim.Data.SymPrim
   ( SymBool,
