--- src/Grisette/IR/SymPrim/Data/Prim/InternedTerm/TermSubstitution.hs-boot
@@ -1,20 +1,0 @@
-{-# LANGUAGE RankNTypes #-}
-
-module Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermSubstitution
-  ( substTerm,
-  )
-where
-
-import {-# SOURCE #-} Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( SupportedPrim,
-    Term,
-    TypedSymbol,
-  )
-
-substTerm ::
-  forall a b.
-  (SupportedPrim a, SupportedPrim b) =>
-  TypedSymbol a ->
-  Term a ->
-  Term b ->
-  Term b
