+++ src/Grisette/IR/SymPrim/Data/Prim/InternedTerm/TermUtils.hs
@@ -13,12 +13,8 @@
 -- Stability   :   Experimental
 -- Portability :   GHC only
 module Grisette.IR.SymPrim.Data.Prim.InternedTerm.TermUtils
-  ( identity,
-    identityWithTypeRep,
-    introSupportedPrimConstraint,
-    extractSymbolicsTerm,
+  ( extractSymbolicsTerm,
     castTerm,
-    pformat,
     someTermsSize,
     someTermSize,
     termSize,
@@ -36,26 +32,22 @@ import Control.Monad.State
   )
 import qualified Data.HashMap.Strict as M
 import qualified Data.HashSet as S
-import Data.Interned (Id)
 import Data.Typeable
-  ( Proxy (Proxy),
-    TypeRep,
-    Typeable,
+  ( Typeable,
     cast,
-    typeRep,
   )
 import Grisette.IR.SymPrim.Data.Prim.InternedTerm.SomeTerm
   ( SomeTerm (SomeTerm),
   )
 import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
-  ( BinaryOp (pformatBinary),
-    SomeTypedSymbol (SomeTypedSymbol),
-    SupportedPrim (pformatCon, pformatSym),
+  ( SomeTypedSymbol (SomeTypedSymbol),
+    SupportedPrim,
     Term
       ( AbsNumTerm,
         AddNumTerm,
         AndBitsTerm,
         AndTerm,
+        ApplyTerm,
         BVConcatTerm,
         BVExtendTerm,
         BVSelectTerm,
@@ -65,7 +57,6 @@ import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
         DivBoundedIntegralTerm,
         DivIntegralTerm,
         EqvTerm,
-        GeneralFunApplyTerm,
         ITETerm,
         LENumTerm,
         LTNumTerm,
@@ -93,101 +84,10 @@ import Grisette.IR.SymPrim.Data.Prim.InternedTerm.Term
         UnaryTerm,
         XorBitsTerm
       ),
-    TernaryOp (pformatTernary),
     TypedSymbol,
-    UnaryOp (pformatUnary),
+    introSupportedPrimConstraint,
   )
 import qualified Type.Reflection as R
-
-identity :: Term t -> Id
-identity = snd . identityWithTypeRep
-{-# INLINE identity #-}
-
-identityWithTypeRep :: forall t. Term t -> (TypeRep, Id)
-identityWithTypeRep (ConTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (SymTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (UnaryTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (BinaryTerm i _ _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (TernaryTerm i _ _ _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (NotTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (OrTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (AndTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (EqvTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ITETerm i _ _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (AddNumTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (UMinusNumTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (TimesNumTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (AbsNumTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (SignumNumTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (LTNumTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (LENumTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (AndBitsTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (OrBitsTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (XorBitsTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ComplementBitsTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ShiftLeftTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ShiftRightTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (RotateLeftTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (RotateRightTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ToSignedTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ToUnsignedTerm i _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (BVConcatTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (BVSelectTerm i _ _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (BVExtendTerm i _ _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (TabularFunApplyTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (GeneralFunApplyTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (DivIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ModIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (QuotIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (RemIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (DivBoundedIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (ModBoundedIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (QuotBoundedIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-identityWithTypeRep (RemBoundedIntegralTerm i _ _) = (typeRep (Proxy @t), i)
-{-# INLINE identityWithTypeRep #-}
-
-introSupportedPrimConstraint :: forall t a. Term t -> ((SupportedPrim t) => a) -> a
-introSupportedPrimConstraint ConTerm {} x = x
-introSupportedPrimConstraint SymTerm {} x = x
-introSupportedPrimConstraint UnaryTerm {} x = x
-introSupportedPrimConstraint BinaryTerm {} x = x
-introSupportedPrimConstraint TernaryTerm {} x = x
-introSupportedPrimConstraint NotTerm {} x = x
-introSupportedPrimConstraint OrTerm {} x = x
-introSupportedPrimConstraint AndTerm {} x = x
-introSupportedPrimConstraint EqvTerm {} x = x
-introSupportedPrimConstraint ITETerm {} x = x
-introSupportedPrimConstraint AddNumTerm {} x = x
-introSupportedPrimConstraint UMinusNumTerm {} x = x
-introSupportedPrimConstraint TimesNumTerm {} x = x
-introSupportedPrimConstraint AbsNumTerm {} x = x
-introSupportedPrimConstraint SignumNumTerm {} x = x
-introSupportedPrimConstraint LTNumTerm {} x = x
-introSupportedPrimConstraint LENumTerm {} x = x
-introSupportedPrimConstraint AndBitsTerm {} x = x
-introSupportedPrimConstraint OrBitsTerm {} x = x
-introSupportedPrimConstraint XorBitsTerm {} x = x
-introSupportedPrimConstraint ComplementBitsTerm {} x = x
-introSupportedPrimConstraint ShiftLeftTerm {} x = x
-introSupportedPrimConstraint RotateLeftTerm {} x = x
-introSupportedPrimConstraint ShiftRightTerm {} x = x
-introSupportedPrimConstraint RotateRightTerm {} x = x
-introSupportedPrimConstraint ToSignedTerm {} x = x
-introSupportedPrimConstraint ToUnsignedTerm {} x = x
-introSupportedPrimConstraint BVConcatTerm {} x = x
-introSupportedPrimConstraint BVSelectTerm {} x = x
-introSupportedPrimConstraint BVExtendTerm {} x = x
-introSupportedPrimConstraint TabularFunApplyTerm {} x = x
-introSupportedPrimConstraint GeneralFunApplyTerm {} x = x
-introSupportedPrimConstraint DivIntegralTerm {} x = x
-introSupportedPrimConstraint ModIntegralTerm {} x = x
-introSupportedPrimConstraint QuotIntegralTerm {} x = x
-introSupportedPrimConstraint RemIntegralTerm {} x = x
-introSupportedPrimConstraint DivBoundedIntegralTerm {} x = x
-introSupportedPrimConstraint ModBoundedIntegralTerm {} x = x
-introSupportedPrimConstraint QuotBoundedIntegralTerm {} x = x
-introSupportedPrimConstraint RemBoundedIntegralTerm {} x = x
-{-# INLINE introSupportedPrimConstraint #-}
 
 extractSymbolicsSomeTerm :: SomeTerm -> S.HashSet SomeTypedSymbol
 extractSymbolicsSomeTerm t1 = evalState (gocached t1) M.empty
@@ -234,7 +134,7 @@ extractSymbolicsSomeTerm t1 = evalState (gocached t1) M.empty
     go (SomeTerm (BVSelectTerm _ _ _ arg)) = goUnary arg
     go (SomeTerm (BVExtendTerm _ _ _ arg)) = goUnary arg
     go (SomeTerm (TabularFunApplyTerm _ func arg)) = goBinary func arg
-    go (SomeTerm (GeneralFunApplyTerm _ func arg)) = goBinary func arg
+    go (SomeTerm (ApplyTerm _ func arg)) = goBinary func arg
     go (SomeTerm (DivIntegralTerm _ arg1 arg2)) = goBinary arg1 arg2
     go (SomeTerm (ModIntegralTerm _ arg1 arg2)) = goBinary arg1 arg2
     go (SomeTerm (QuotIntegralTerm _ arg1 arg2)) = goBinary arg1 arg2
@@ -291,7 +191,7 @@ castTerm t@BVConcatTerm {} = cast t
 castTerm t@BVSelectTerm {} = cast t
 castTerm t@BVExtendTerm {} = cast t
 castTerm t@TabularFunApplyTerm {} = cast t
-castTerm t@GeneralFunApplyTerm {} = cast t
+castTerm t@ApplyTerm {} = cast t
 castTerm t@DivIntegralTerm {} = cast t
 castTerm t@ModIntegralTerm {} = cast t
 castTerm t@QuotIntegralTerm {} = cast t
@@ -301,50 +201,6 @@ castTerm t@ModBoundedIntegralTerm {} = cast t
 castTerm t@QuotBoundedIntegralTerm {} = cast t
 castTerm t@RemBoundedIntegralTerm {} = cast t
 {-# INLINE castTerm #-}
-
-pformat :: forall t. (SupportedPrim t) => Term t -> String
-pformat (ConTerm _ t) = pformatCon t
-pformat (SymTerm _ sym) = pformatSym sym
-pformat (UnaryTerm _ tag arg1) = pformatUnary tag arg1
-pformat (BinaryTerm _ tag arg1 arg2) = pformatBinary tag arg1 arg2
-pformat (TernaryTerm _ tag arg1 arg2 arg3) = pformatTernary tag arg1 arg2 arg3
-pformat (NotTerm _ arg) = "(! " ++ pformat arg ++ ")"
-pformat (OrTerm _ arg1 arg2) = "(|| " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (AndTerm _ arg1 arg2) = "(&& " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (EqvTerm _ arg1 arg2) = "(= " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (ITETerm _ cond arg1 arg2) = "(ite " ++ pformat cond ++ " " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (AddNumTerm _ arg1 arg2) = "(+ " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (UMinusNumTerm _ arg) = "(- " ++ pformat arg ++ ")"
-pformat (TimesNumTerm _ arg1 arg2) = "(* " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (AbsNumTerm _ arg) = "(abs " ++ pformat arg ++ ")"
-pformat (SignumNumTerm _ arg) = "(signum " ++ pformat arg ++ ")"
-pformat (LTNumTerm _ arg1 arg2) = "(< " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (LENumTerm _ arg1 arg2) = "(<= " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (AndBitsTerm _ arg1 arg2) = "(& " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (OrBitsTerm _ arg1 arg2) = "(| " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (XorBitsTerm _ arg1 arg2) = "(^ " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (ComplementBitsTerm _ arg) = "(~ " ++ pformat arg ++ ")"
-pformat (ShiftLeftTerm _ arg n) = "(shl " ++ pformat arg ++ " " ++ pformat n ++ ")"
-pformat (ShiftRightTerm _ arg n) = "(shr " ++ pformat arg ++ " " ++ pformat n ++ ")"
-pformat (RotateLeftTerm _ arg n) = "(rotl " ++ pformat arg ++ " " ++ pformat n ++ ")"
-pformat (RotateRightTerm _ arg n) = "(rotr " ++ pformat arg ++ " " ++ pformat n ++ ")"
-pformat (ToSignedTerm _ arg) = "(u2s " ++ pformat arg ++ " " ++ ")"
-pformat (ToUnsignedTerm _ arg) = "(s2u " ++ pformat arg ++ " " ++ ")"
-pformat (BVConcatTerm _ arg1 arg2) = "(bvconcat " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (BVSelectTerm _ ix w arg) = "(bvselect " ++ show ix ++ " " ++ show w ++ " " ++ pformat arg ++ ")"
-pformat (BVExtendTerm _ signed n arg) =
-  (if signed then "(bvsext " else "(bvzext ") ++ show n ++ " " ++ pformat arg ++ ")"
-pformat (TabularFunApplyTerm _ func arg) = "(apply " ++ pformat func ++ " " ++ pformat arg ++ ")"
-pformat (GeneralFunApplyTerm _ func arg) = "(apply " ++ pformat func ++ " " ++ pformat arg ++ ")"
-pformat (DivIntegralTerm _ arg1 arg2) = "(div " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (ModIntegralTerm _ arg1 arg2) = "(mod " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (QuotIntegralTerm _ arg1 arg2) = "(quot " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (RemIntegralTerm _ arg1 arg2) = "(rem " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (DivBoundedIntegralTerm _ arg1 arg2) = "(div " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (ModBoundedIntegralTerm _ arg1 arg2) = "(mod " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (QuotBoundedIntegralTerm _ arg1 arg2) = "(quot " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-pformat (RemBoundedIntegralTerm _ arg1 arg2) = "(rem " ++ pformat arg1 ++ " " ++ pformat arg2 ++ ")"
-{-# INLINE pformat #-}
 
 someTermsSize :: [SomeTerm] -> Int
 someTermsSize terms = S.size $ execState (traverse goSome terms) S.empty
@@ -385,7 +241,7 @@ someTermsSize terms = S.size $ execState (traverse goSome terms) S.empty
     go t@(BVSelectTerm _ _ _ arg) = goUnary t arg
     go t@(BVExtendTerm _ _ _ arg) = goUnary t arg
     go t@(TabularFunApplyTerm _ func arg) = goBinary t func arg
-    go t@(GeneralFunApplyTerm _ func arg) = goBinary t func arg
+    go t@(ApplyTerm _ func arg) = goBinary t func arg
     go t@(DivIntegralTerm _ arg1 arg2) = goBinary t arg1 arg2
     go t@(ModIntegralTerm _ arg1 arg2) = goBinary t arg1 arg2
     go t@(QuotIntegralTerm _ arg1 arg2) = goBinary t arg1 arg2
